Add header for sc_lib
